// src/app/config.enum.ts
export enum ConfigDataKeys {
  baseURL = 'baseURL',
  authorization = 'authorization',
  url = 'url',
  workspaceName = 'workspaceName',
  layers = 'layers',
  geoserver = 'geoserver',
  geomocusServer = 'geomocusServer',
  createWorkspace = 'createWorkspace',
  localhost = 'localhost',
  connectionParameters = 'connectionParameters',
  save = 'save',
  bookmarkurl = 'bookmarkurl',
  appurl = 'appurl',
  layersave = 'layersave',
  layerList = 'layerList'
}